+ Move flag anisoOnMesh from ACov to CorAniso
